make hook to also return the callback
